requirements.txt
certifi==2024.12.14
influxdb3-python==0.10.0
ping3==4.0.8
pyarrow==19.0.0
python-dateutil==2.9.0.post0
reactivex==4.0.4
six==1.17.0
speedtest-cli==2.1.3
typing_extensions==4.12.2
urllib3==2.3.0
